requirements.txt
requests
pandas
yfinance
lxml
html5lib
beautifulsoup4
python-dateutil
google-generativeai
python-dotenv
